+++ example/lib/features/edit_keyvalue.dart
@@ -31,7 +31,8 @@ import 'package:demopod/constants/app.dart';
 import 'package:demopod/dialogs/alert.dart';
 import 'package:demopod/utils/rdf.dart';
 import 'package:editable/editable.dart';
-import 'package:solidui/solidui.dart' show SolidScaffold, getKeyFromUserIfRequired;
+import 'package:solidui/solidui.dart'
+    show SolidScaffold, getKeyFromUserIfRequired;
 
 import 'package:solidpod/solidpod.dart' show isUserLoggedIn, writePod;
